+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "memcached-typed",
-  "version": "3.1.0",
+  "version": "4.0.0",
   "description": "Typescript Memcached Wrapper, with Promise and Template helper",
   "main": "./dst/index.js",
   "typings": "./dst/index.d.ts",
@@ -11,7 +11,7 @@
     "prepublish": "npm run build",
     "pretest": "npm run build",
     "ci:publish": "publish",
-    "test": "MEMCACHED_URL=127.0.0.1 mocha dst/**/__test__/**/*.js"
+    "test": "MEMCACHED_URL=127.0.0.1 REDIS_URL=redis://127.0.0.1/3 mocha dst/**/__test__/**/*.js"
   },
   "repository": {
     "type": "git",
@@ -26,6 +26,7 @@
   "devDependencies": {
     "@types/chai": "^4.1.3",
     "@types/chai-as-promised": "^7.1.0",
+    "@types/memcached": "^2.2.5",
     "@types/mocha": "^2.2.48",
     "@types/node": "^6.0.106",
     "chai": "^4.1.2",
@@ -35,7 +36,10 @@
     "typescript": "^2.4.2"
   },
   "dependencies": {
-    "@types/memcached": "^2.2.5",
-    "memcached": "^2.2.2"
+    "@types/bluebird": "^3.5.20",
+    "@types/redis": "^2.8.6",
+    "bluebird": "^3.5.1",
+    "memcached": "^2.2.2",
+    "redis": "^2.8.0"
   }
 }
